Put vocabylary tests into same module.
Move TestCommitteeMemberVocabulary to the same module as the other vocabulary
tests.
